+++ core/src/components/modal/test/test.utils.ts
@@ -1,5 +1,24 @@
+import type { E2EPage } from '@stencil/core/testing';
 import { newE2EPage } from '@stencil/core/testing';
 import { generateE2EUrl } from '@utils/test';
+
+export const openModal = async (page: E2EPage, selector: string) => {
+  const ionModalWillPresent = await page.spyOnEvent('ionModalWillPresent');
+  const ionModalDidPresent = await page.spyOnEvent('ionModalDidPresent');
+
+  await page.click(selector);
+
+  await ionModalWillPresent.next();
+  await ionModalDidPresent.next();
+
+  await page.waitForSelector(selector);
+
+  const modal = await page.find('ion-modal');
+  await modal.waitForVisible();
+  await page.waitForTimeout(100);
+
+  return modal;
+};
 
 export const testModal = async (type: string, selector: string, expectUnmount = true, rtl = false) => {
   const pageUrl = generateE2EUrl('modal', type, rtl);
@@ -9,21 +28,10 @@ export const testModal = async (type: string, selector: string, expectUnmount = 
   });
 
   const screenshotCompares = [];
-  const ionModalWillPresent = await page.spyOnEvent('ionModalWillPresent');
-  const ionModalDidPresent = await page.spyOnEvent('ionModalDidPresent');
   const ionModalWillDismiss = await page.spyOnEvent('ionModalWillDismiss');
   const ionModalDidDismiss = await page.spyOnEvent('ionModalDidDismiss');
 
-  await page.click(selector);
-
-  await ionModalWillPresent.next();
-  await ionModalDidPresent.next();
-
-  await page.waitForSelector(selector);
-
-  let modal = await page.find('ion-modal');
-  await modal.waitForVisible();
-  await page.waitForTimeout(100);
+  let modal = await openModal(page, selector);
 
   screenshotCompares.push(await page.compareScreenshot());