+++ core/src/components/reorder-group/test/interactive/e2e.ts
@@ -1,4 +1,4 @@
-import type * as pd from '@stencil/core/dist/testing/puppeteer/puppeteer-declarations';
+import type { E2EPage } from '@stencil/core/testing';
 import { newE2EPage } from '@stencil/core/testing';
 import { getElementProperty, queryDeep } from '@utils/test';
 
@@ -53,7 +53,7 @@ test('reorder: interactive', async () => {
 
 const moveItem = async (
   id: string,
-  page: pd.E2EPage,
+  page: E2EPage,
   direction: 'up' | 'down' = 'up',
   numberOfSpaces = 1,
   ...parentSelectors: string[]